+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added ascii sdf file format
 - Added ignore_errors kwarg to Batch.execute to ignore errors during parameter calculation and fill the respective 
   values with nan instead
+- Added writing of ascii and binary sdf file format
 ## v0.9.2
 - Fixed bug with type hinting
 ## v0.9.1